[FrameworkBundle] Move AbstractController::getParameter() from the trait to the class & use PSR-11
